C# and Python codelenses draw with correct criteria
