+++ src/shared/codelens/codeLensUtils.ts
@@ -3,6 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import * as path from 'path'
 import * as vscode from 'vscode'
 
 import { detectLocalTemplates } from '../../lambda/local/detectLocalTemplates'
@@ -19,6 +20,8 @@ import { ExtContext } from '../extensions'
 import { recordLambdaInvokeLocal, Result, Runtime } from '../telemetry/telemetry'
 import { TypescriptLambdaHandlerSearch } from '../typescriptLambdaHandlerSearch'
 import { nodeJsRuntimes } from '../../lambda/models/samLambdaRuntime'
+import { dirnameWithTrailingSlash } from '../utilities/pathUtils'
+import { isInDirectory } from '../filesystemUtilities'
 
 export type Language = 'python' | 'javascript' | 'csharp'
 
@@ -351,4 +354,37 @@ export function initializeTypescriptCodelens(context: ExtContext): void {
             }
         })
     )
+}
+
+export async function findParentProjectFile(
+    sourceCodeUri: vscode.Uri,
+    searchPattern: string,
+    findWorkspaceFiles: typeof vscode.workspace.findFiles = vscode.workspace.findFiles
+): Promise<vscode.Uri | undefined> {
+    const workspaceFolder = vscode.workspace.getWorkspaceFolder(sourceCodeUri)
+    if (!workspaceFolder) {
+        return undefined
+    }
+
+    const workspaceProjectFiles: vscode.Uri[] = await findWorkspaceFiles(
+        new vscode.RelativePattern(workspaceFolder, searchPattern)
+    )
+
+    // Use the project file "closest" in the parent chain to sourceCodeUri
+    // Assumption: only one .csproj file will exist in a given folder
+    let parentProjectFiles = workspaceProjectFiles
+        .filter(uri => {
+            const dirname = dirnameWithTrailingSlash(uri.fsPath)
+
+            return sourceCodeUri.fsPath.startsWith(dirname)
+        })
+        .sort((a, b) => {
+            if (isInDirectory(path.parse(a.fsPath).dir, path.parse(b.fsPath).dir)) {
+                return 1
+            }
+
+            return -1
+        })
+
+    return parentProjectFiles.length === 0 ? undefined : parentProjectFiles[0]
 }
