+++ src/shared/codelens/pythonCodeLensProvider.ts
@@ -28,6 +28,7 @@ import { PythonDebugAdapterHeartbeat } from './pythonDebugAdapterHeartbeat'
 import { getLocalRootVariants } from '../utilities/pathUtils'
 import { SamLaunchRequestArgs } from '../sam/debugger/samDebugSession'
 import * as pathutil from '../../shared/utilities/pathUtils'
+import { findParentProjectFile } from './codeLensUtils'
 
 const PYTHON_DEBUG_ADAPTER_RETRY_DELAY_MS = 1000
 export const PYTHON_LANGUAGE = 'python'
@@ -44,7 +45,16 @@ export async function getSamProjectDirPathForFile(filepath: string): Promise<str
 }
 
 export async function getLambdaHandlerCandidates(uri: vscode.Uri): Promise<LambdaHandlerCandidate[]> {
+    const requirementsFile = await findParentProjectFile(uri, path.join('**', 'requirements.txt'))
+    if (!requirementsFile) {
+        return []
+    }
     const filename = uri.fsPath
+    const parsedPath = path.parse(filename)
+    const handlerPath = path
+        .relative(path.parse(requirementsFile.fsPath).dir, path.join(parsedPath.dir, parsedPath.name))
+        .split(path.sep)
+        .join('.')
 
     const symbols: vscode.DocumentSymbol[] =
         (await vscode.commands.executeCommand<vscode.DocumentSymbol[]>('vscode.executeDocumentSymbolProvider', uri)) ??
@@ -53,7 +63,7 @@ export async function getLambdaHandlerCandidates(uri: vscode.Uri): Promise<Lambd
     return symbols.filter(isTopLevelFunction).map<LambdaHandlerCandidate>(symbol => {
         return {
             filename,
-            handlerName: `${path.parse(filename).name}.${symbol.name}`,
+            handlerName: `${handlerPath}${path.parse(filename).name}.${symbol.name}`,
             range: symbol.range,
         }
     })
